fix: better handle transactions
