+++ cli/Cargo.toml
@@ -20,4 +20,5 @@ anyhow.workspace = true
 clap = { workspace = true, features = ["derive", "env"] }
 comfy-table = "7.2.1"
 dotenv.workspace = true
+sqlx = { workspace = true, features = ["postgres"] }
 tokio.workspace = true
