+++ validity/Cargo.toml
@@ -47,7 +47,7 @@ reqwest.workspace = true
 bincode.workspace = true
 base64.workspace = true
 serde_repr = "0.1.19"
-sqlx = { version = "0.8.3", features = [
+sqlx = { workspace = true, features = [
     "postgres",
     "chrono",
     "bigdecimal",
