+++ validity/src/db/mod.rs
@@ -1,5 +1,10 @@
 mod client;
+mod queries;
 mod types;
+
+pub use queries::*;
 
 // Re-export everything to maintain the current API
 pub use types::*;
+
+pub use client::init_db;
